refactor: reduce app state coupling
